Fix MSVC warnings in gnm and gcore
git-svn-id: https://svn.osgeo.org/gdal/trunk@31627 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/gcore/gdalcolortable.cpp
@@ -297,7 +297,7 @@ GDALColorTableH CPL_STDCALL GDALCloneColorTable( GDALColorTableH hTable )
 int GDALColorTable::GetColorEntryCount() const
 
 {
-    return aoEntries.size();
+    return static_cast<int>(aoEntries.size());
 }
 
 /************************************************************************/
